A couple small fixed but mostly removing whisperx for fasterwhsiper
