+++ EXAMPLE.config
@@ -1,6 +1,7 @@
 
 ## MAKE A COPY OF THIS CALLED config.py
-VERBOSE = False
+VERBOSE = True
+USE_GPU = False  # Set to True to use GPU acceleration. Refer to the README for instructions on installing PyTorch with CUDA support.
 
 
 ### COMPLETIONS API SETTINGS  ###
@@ -36,21 +37,21 @@ COMPLETION_MODEL = "gpt-4-0125-preview"
 
 ### Transcription API Settings ###
 
-## Whisper X local transcription API EXAMPLE ##
-TRANSCRIPTION_API = "whisperx" #local transcription!
-WHISPER_MODEL = "tiny" # (tiny, base, small, medium, large) Turn this up to "base" if the transcription is too bad
-TRANSCRIPTION_LANGUAGE = "en" 
-WHISPER_BATCH_SIZE = 16
-WHISPER_MODEL_PATH = None # you can point this to an existing model or leave it set to none
+## Faster Whisper local transcription ###
+# TRANSCRIPTION_API = "FasterWhisper" # this will use the local whisper model
+# WHISPER_MODEL = "tiny.en" # If you prefer not to use english set it to "tiny", if the transcription quality is too low then set it to "base" but this will be a little slower
+# BEAM_SIZE = 5
 
-## Transcription API Settings ###
-# TRANSCRIPTION_API = "openai" # this will use the hosted openai api
+
+## OPENAI Hosted Transcription ###
+TRANSCRIPTION_API = "openai" # this will use the hosted openai api
 
 
 
 ### TTS SETTINGS ###
+TTS_ENGINE="openai" # 'piper' or 'openai' piper is local and fast but openai is better sounding
+
 PIPER_VOICE = "default_female_voice"
-TTS_ENGINE="piper" # 'piper' or 'openai' piper is local and fast but openai is better sounding
 OPENAI_VOICE = "nova"
 
 ### PROMPTS ###
