+++ faster_whisper_requirements.txt
@@ -1,0 +1,1 @@
+faster-whisper
