--- local_whisper_requirements.txt
@@ -1,4 +1,0 @@
-torch
-torchvision
-torchaudio
-git+https://github.com/m-bain/whisperx.git
